# app/Http/Controllers/ProductCategoryController.php
<?php

namespace App\Http\Controllers;

use App\Http\Requests\StoreProductCategoryRequest;
use App\Http\Resources\ProductCategoryResource;
use App\Models\ProductCategory;
use App\Services\ProductCategoryService;
use Illuminate\Http\Request;

class ProductCategoryController extends Controller
{
    protected $productCategory;


    public function __construct(ProductCategoryService $productCategory)
    {
        $this->productCategory = $productCategory;
    }

    /**
     * @OA\PathItem(
     *  title= "My first API",
     *  version = "0.1"
     * )
     *
     * **/
    public function index()
    {
        $product = Product::with('ProductCategory')->get();

        $productList = ProductResource::collection($product);

        return $this->success($productList, 'success', 200);
    }

     /**

 * @OA\Post(
 *     path="/api/v1/product-categories",
 *     summary="Post all product-categories",
 *     operationId="postProductCategory",
 *     tags={"ProductCategory"},
 *      @OA\Parameter( name="ProductCategoryCode", in="query", description="ProductCategory Name",required=true, @OA\Schema(type="string")),
 *      @OA\Parameter( name="ProductCategoryName", in="query", description="mobile No",required=true, @OA\Schema(type="string")),
 *     @OA\Response(response=200, description="Successful operation",  @OA\JsonContent() ),
 *     security={{"bearerAuth":{}}}
 * )
 */
    public function store(StoreProductCategoryRequest $request)
    {

        $productCategory = $this->productCategory->insert($request->all());

        if ($productCategory) {
            return response()->json([
                'data' => $productCategory,
                'status' => true
            ], 200);
        }
    }

     /**
 * @OA\Get(
 *     path="/api/v1/product-categories/{id}",
 *     summary="Show product-categories",
 *     operationId="showProductCategories",
 *     tags={"ProductCategory"},
 *     @OA\Parameter( name="id", in="path", description="ID of the productCategory member", required=true,
 *       @OA\Schema(
 *             type="integer",
 *             format="int64"
 *         )
 *     ),
 *@OA\Response(  response=200, description="Successful operation",  @OA\JsonContent()  ),
 *    security={{"bearerAuth":{}}}
 * )
 */
    public function show(string $id)
    {
        $productCategory = $this->productCategory->getDataById($id);

        if ($productCategory) {
            return response()->json([
                'data' => $productCategory,
                'status' => true
            ], 200);
        } else {
            return response()->json([
                'message' => 'No data found',
                'status' => false
            ], 404);
        }
    }

       /**

 * @OA\Put(
 *     path="/api/v1/product-categories/{id}",
 *     summary="update all product-categories",
 *     operationId="updateProductCategory",
 *     tags={"ProductCategory"},
 *     @OA\Parameter( name="ProductCategoryId", in="path", description="Enter Id you want to update", required=true,
 *       @OA\Schema(
 *             type="integer",
 *             format="int64"
 *         )
 *     ),
 *      @OA\Parameter( name="ProductCategoryCode", in="query", description="ProductCategory Name",required=false, @OA\Schema(type="string")),
 *      @OA\Parameter( name="ProductCategoryName", in="query", description="mobile No",required=false, @OA\Schema(type="string")),

 *     @OA\Response(response=200, description="Successful operation",  @OA\JsonContent() ),
 *     security={{"bearerAuth":{}}}
 * )
 */
    public function update(Request $request, string $id)
    {
        $productCategory = $this->productCategory->update($request->all(), $id);

        if ($productCategory) {
            return response()->json([
                'message' => 'Successfully updated data',
                'status' => true
            ], 200);
        } else {
            return response()->json([
                'message' => 'No data found',
                'status' => false
            ], 404);
        }
    }

        /**

    * @OA\Delete(
    *     path="/api/v1/product-categories/{id}",
    *     summary="delete product-categories",
    *     operationId="deleteProductCategory",
    *     tags={"ProductCategory"},
    *     @OA\Parameter( name="id", in="path", description="Enter Id you want to delete", required=true,
    *       @OA\Schema(
    *             type="integer",
    *             format="int64"
    *         )
    *     ),
    *      @OA\Response(  response=200, description="Successful operation",  @OA\JsonContent()  ),

    *     security={{"bearerAuth":{}}}
    * )
    */   
    public function destroy(string $id)
    {
        $productCategory = $this->productCategory->destroy($id);

        if ($productCategory) {
            return response()->json([
                'message' => 'Successfully deleted data',
                'status' => true
            ], 200);
        } else {
            return response()->json([
                'message' => 'No data found',
                'status' => false
            ], 404);
        }
    }
}