# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger with Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/product-categories": {
            "post": {
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Post all product-categories",
                "operationId": "postProductCategory",
                "parameters": [
                    {
                        "name": "ProductCategoryCode",
                        "in": "query",
                        "description": "ProductCategory Name",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ProductCategoryName",
                        "in": "query",
                        "description": "mobile No",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/product-categories/{id}": {
            "get": {
                "tags": [
                    "ProductCategory"
                ],
                "summary": "Show product-categories",
                "operationId": "showProductCategories",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the productCategory member",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "ProductCategory"
                ],
                "summary": "update all product-categories",
                "operationId": "updateProductCategory",
                "parameters": [
                    {
                        "name": "ProductCategoryId",
                        "in": "path",
                        "description": "Enter Id you want to update",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    },
                    {
                        "name": "ProductCategoryCode",
                        "in": "query",
                        "description": "ProductCategory Name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "ProductCategoryName",
                        "in": "query",
                        "description": "mobile No",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "ProductCategory"
                ],
                "summary": "delete product-categories",
                "operationId": "deleteProductCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Enter Id you want to delete",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}